@@ -1,7 +1,7 @@
 {
   "name": "chainner",
   "productName": "chaiNNer",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "description": "A flowchart based image processing GUI",
   "main": ".webpack/main",
   "scripts": {
@@ -9,6 +9,9 @@
     "dev": "concurrently \"nodemon ./backend/run.py 8000\" \"electron-forge start -- --no-backend\"",
     "package": "electron-forge package",
     "make": "electron-forge make",
+    "make-linux-zip": "electron-forge make --targets @electron-forge/maker-zip --platform linux",
+    "make-win-zip": "electron-forge make --targets @electron-forge/maker-zip --platform win32",
+    "make-mac-zip": "electron-forge make --targets @electron-forge/maker-zip --platform darwin",
     "publish": "electron-forge publish",
     "lint": "eslint . --fix"
   },
@@ -118,17 +121,20 @@
     "concurrently": "^6.5.1",
     "css-loader": "^6.2.0",
     "electron": "^17.0.0",
+    "electron-installer-common": "^0.10.3",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
+    "file-loader": "^6.2.0",
+    "image-webpack-loader": "^8.1.0",
     "node-loader": "^2.0.0",
     "semver-regex": ">=3.1.3",
     "style-loader": "^3.2.1"
   },
   "dependencies": {
     "@chakra-ui/icons": "^1.0.15",
-    "@chakra-ui/react": "^1.6.8",
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
+    "@chakra-ui/react": "^1.8.6",
+    "@emotion/react": "^11.8.1",
+    "@emotion/styled": "^11.8.1",
     "@fontsource/open-sans": "^4.5.1",
     "@geoffcox/react-splitter": "^2.1.1",
     "@react-hook/window-size": "^3.0.7",
@@ -136,31 +142,21 @@
     "decompress": "^4.2.1",
     "electron-log": "^4.4.1",
     "electron-squirrel-startup": "^1.0.0",
-    "file-loader": "^6.2.0",
-    "framer-motion": "^4.1.17",
-    "hasbin": "^1.2.3",
+    "framer-motion": "^6.2.8",
     "html-to-image": "^1.9.0",
     "image-js": "^0.33.0",
-    "image-webpack-loader": "^8.1.0",
-    "lodash": "^4.17.21",
-    "meow": "^10.1.1",
     "nodejs-file-downloader": "^4.9.2",
     "os-utils": "^0.0.14",
-    "pill-pity": "^0.2.2",
     "portfinder": "^1.0.28",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0-rc.1",
+    "react-dom": "^18.0.0-rc.1",
     "react-flow-renderer": "^9.6.8",
-    "react-hotkeys": "^2.0.0",
     "react-hotkeys-hook": "^3.4.4",
     "react-icons": "^4.2.0",
-    "react-resize-aware": "^3.1.1",
-    "react-use-websocket": "^2.9.1",
     "semver": "^7.3.5",
     "systeminformation": "^5.9.15",
-    "tinycolor2": "^1.4.2",
     "use-debounce": "^7.0.1",
     "use-http": "^1.0.26",
-    "uuid": "^3.4.0"
+    "uuid": "^8.3.2"
   }
 }
